Remove the flags package
Refactor cmd/server
