+++ cmd/tmixer/options/options.go
@@ -1,33 +1,15 @@
 package options
 
 import (
-	"samuellando.com/tmixer/internal/config"
-	"samuellando.com/tmixer/internal/flags"
+	"flag"
 )
 
-func ptr[T any](v T) *T {
-	return &v
-}
+var FLAG_SET = flag.NewFlagSet("", flag.ContinueOnError)
 
-var FLAGS = map[string]flags.Flag{
-	"help": {
-		ShortName:   "h",
-		Description: "display a help message to stdout",
-		Usage:       "--help or -h",
-		Bare:        true,
-		ParseInput: func(s string, c *config.Config) error {
-			c.DisplayHelp = ptr(true)
-			return nil
-		},
-	},
-	"verbose": {
-		ShortName:   "v",
-		Description: "display information about the run",
-		Usage:       "--verbose or -v",
-		Bare:        true,
-		ParseInput: func(s string, c *config.Config) error {
-			c.DisplayLog = ptr(true)
-			return nil
-		},
-	},
+var DISPLAY_HELP bool
+var VERBOSE bool
+
+func init() {
+	FLAG_SET.BoolVar(&DISPLAY_HELP, "help", false, "Display the help message")
+	FLAG_SET.BoolVar(&VERBOSE, "verbose", false, "Display debug logs")
 }
